+++ examples/integrations/Ninetailed/sample-apps/app-using-v1.1-cdn/src/App.js
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 import { Ninetailed } from '@ninetailed/experience.js';
 import logo from './logo.svg';
 import './App.css';
-import { CDN_URL, CONFIG_URL, DATAPLANE_URL, WRITE_KEY } from './config';
+import { DATAPLANE_URL, WRITE_KEY } from './config';
 
 export const ninetailed = new Ninetailed({
   // REQUIRED. An API key uniquely identifying your Ninetailed account.
@@ -50,12 +50,12 @@ function App() {
       let r = document.createElement('script');
       r.type = 'text/javascript';
       r.async = true;
-      r.src = CDN_URL;
+      r.src = 'https://cdn.rudderlabs.com/v1.1/rudder-analytics.min.js';
       let a = document.getElementsByTagName('script')[0];
       a.parentNode.insertBefore(r, a);
     };
     e.loadJS();
-    e.load(WRITE_KEY, DATAPLANE_URL, { configUrl: CONFIG_URL });
+    e.load(WRITE_KEY, DATAPLANE_URL);
     /* eslint-disable */
   }, []);
   const page = () => {