+++ packages/analytics-js-common/src/constants/integrations/Heap/constants.ts
@@ -1,6 +1,6 @@
+import { HEAP_NAME as NAME, HEAP_DISPLAY_NAME as DISPLAY_NAME } from '../Destinations';
+
 const DIR_NAME = 'Heap';
-const NAME = 'HEAP';
-const DISPLAY_NAME = 'Heap.io';
 
 const DISPLAY_NAME_TO_DIR_NAME_MAP = { [DISPLAY_NAME]: DIR_NAME };
 const CNameMapping = {