+++ packages/analytics-js-common/src/constants/integrations/LaunchDarkly/constants.ts
@@ -1,6 +1,9 @@
+import {
+  LAUNCHDARKLY_NAME as NAME,
+  LAUNCHDARKLY_DISPLAY_NAME as DISPLAY_NAME,
+} from '../Destinations';
+
 const DIR_NAME = 'LaunchDarkly';
-const NAME = 'LAUNCHDARKLY';
-const DISPLAY_NAME = 'LaunchDarkly';
 
 const DISPLAY_NAME_TO_DIR_NAME_MAP = { [DISPLAY_NAME]: DIR_NAME };
 const CNameMapping = {