+++ packages/analytics-js-common/src/constants/integrations/LiveChat/constants.ts
@@ -1,6 +1,6 @@
+import { LIVECHAT_NAME as NAME, LIVECHAT_DISPLAY_NAME as DISPLAY_NAME } from '../Destinations';
+
 const DIR_NAME = 'LiveChat';
-const NAME = 'LIVECHAT';
-const DISPLAY_NAME = 'livechat';
 
 const DISPLAY_NAME_TO_DIR_NAME_MAP = { [DISPLAY_NAME]: DIR_NAME };
 const CNameMapping = {