+++ packages/analytics-js-common/src/constants/integrations/SpotifyPixel/constants.ts
@@ -1,6 +1,9 @@
+import {
+  SPOTIFYPIXEL_NAME as NAME,
+  SPOTIFYPIXEL_DISPLAY_NAME as DISPLAY_NAME,
+} from '../Destinations';
+
 const DIR_NAME = 'SpotifyPixel';
-const NAME = 'SPOTIFYPIXEL';
-const DISPLAY_NAME = 'Spotify Pixel';
 
 const DISPLAY_NAME_TO_DIR_NAME_MAP = { [DISPLAY_NAME]: DIR_NAME };
 const CNameMapping = {
@@ -10,11 +13,4 @@ const CNameMapping = {
   SPOTIFY_PIXEL: NAME,
 };
 
-export {
-  NAME,
-  CNameMapping,
-  DISPLAY_NAME_TO_DIR_NAME_MAP,
-  DISPLAY_NAME,
-  DIR_NAME,
-};
-
+export { NAME, CNameMapping, DISPLAY_NAME_TO_DIR_NAME_MAP, DISPLAY_NAME, DIR_NAME };