+++ packages/analytics-js-integrations/src/integrations/Optimizely/browser.js
@@ -126,7 +126,7 @@ class Optimizely {
           sendCampaignData(campaignState);
         }
       } catch (e) {
-        logger.error('Page loaded without Optimizely.')
+        logger.error('Page loaded without Optimizely.');
       }
     };
 
@@ -140,7 +140,7 @@ class Optimizely {
       if (!state) {
         return undefined;
       }
-      const referrer = state.getRedirectInfo() && state.getRedirectInfo().referrer;
+      const referrer = state.getRedirectInfo()?.referrer;
       if (!referrer) {
         return undefined;
       }
@@ -168,7 +168,7 @@ class Optimizely {
       try {
         state = window?.optimizely?.get('state');
       } catch (e) {
-        state = undefined
+        state = undefined;
       }
       if (state) {
         const referrer = checkReferrer();
@@ -194,7 +194,6 @@ class Optimizely {
           },
         });
       }
-
     };
     registerCurrentlyActiveCampaigns();
     registerFutureActiveCampaigns();