+++ packages/analytics-js-plugins/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.6.21](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.6.20...@rudderstack/analytics-js-plugins@3.6.21) (2025-01-03)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.14.15`
+* `@rudderstack/analytics-js-cookies` updated to version `0.4.18`
 ## [3.6.20](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.6.19...@rudderstack/analytics-js-plugins@3.6.20) (2024-12-17)
 
 ### Dependency Updates